+++ CHANGELOG.md
@@ -5,7 +5,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Feature: Add animation on `List.ButtonItem` hover.
-- Feature: Add new viewport range to MediaQueries HOC.
+- Feature: Add new viewport range to mediaQueries HOC.
+- Feature: Deprecated `viewportIsMobile` and `viewportIsTabletOrLess` on
+  mediaQueries HOC. Please use `viewportIsXS` and `viewportIsMOrLess` instead
+  now.
 
 ## [23.12.0] - 2018-11-15